some kexec MIPS improvements
-  using simple   mips* ) in configure.ac to make it compilable on mips2
   and mips64
- remove kexec/arch/mips/mips-setup-simple.S which prepares cmdline for
  new kernel, it is better to move this work to kernel code. BTW this code was
  compilable only on o32 because of t4 is not defined on 64-64 or n32 MIPS ABIs.
- simple put cmdline as string, kernel code should catch cmdline like this

int board_kexec_prepare(struct kimage *image)
{
    int i;
    char *bootloader = "kexec";
    board_boot_desc_ptr->argc = 0;
    for(i=0;i<image->nr_segments;i++)
    {
        printk("segment %d
        if (!strncmp(bootloader, (char*)image->segment[i].buf,
                strlen(bootloader)))
        {
            /*
             * convert command line string to array
             * of parameters (as bootloader does).
             */
            int argc = 0, offt;
            char *str = (char *)image->segment[i].buf;
            char *ptr = strchr(str, ' ');
            while (ptr && (ARGV_MAX_ARGS > argc)) {
                *ptr = '\0';
                if (ptr[1] != ' ') {
                    offt = (int)(ptr - str + 1);
                    boot_desc_ptr->argv[argc] =
                        image->segment[i].mem + offt;
                    argc++;
                }
                ptr = strchr(ptr + 1, ' ');
            }
            boot_desc_ptr->argc = argc;
            break;
        }
    }
   Keep it as string make code simple and more readable.

- add crashdump support
- do not redefine syscalls numbers if they defined in system
  remove fixups for /proc/iomem. If your board provides wrong /proc/iomem please
  fix kernel, or at least you local version of kexec. No need to support it in
  main line. At least add option --fake-iomem
- some minor fixes

+++ kexec/crashdump.h
@@ -26,7 +26,7 @@ struct crash_elf_info {
 	unsigned long backup_src_start;
 	unsigned long backup_src_end;
 
-	unsigned long page_offset;
+	unsigned long long page_offset;
 	unsigned long lowmem_limit;
 
 	int (*get_note_info)(int cpu, uint64_t *addr, uint64_t *len);
